adding the plan sidebar and nested routes
